pyproject.toml
[tool.poetry]
name = "brainrushbot"
version = "0.1.0"
description = ""
authors = ["marakhd <marakin09@mail.ru>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.18.0"
pydantic-settings = "^2.8.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.23.0"}
asyncpg = "^0.30.0"
aerich = "^0.8.1"

[tool.aerich]
tortoise_orm = "src.config.AERICH_CONFIG"
location = "./src/migrations"
src_folder = "./src"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
